requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
Werkzeug==3.1.3

SQLAlchemy~=2.0.36
Flask-Login~=0.6.3
WTForms~=3.2.1
Flask-Migrate~=4.0.7
python-dotenv~=1.0.1
alembic~=1.14.0